remove needless timestamps
This is an attempt at cache-busting, but it's not really effective due
to needing manual update. Besides, this file hasn't been changed in at
least the last year.

We're currently hosting this file with a max-age of 604800, which is 7
days. We should probably consider lowering that somewhat, but for now
this should work.