Bring back styles (#132)
* add styles back to the repo

These are the CSS files that are currently live at content.pouet.net.
The website depends on their content, and they should be updated in
lockstep, so let's not try to use them from a separate site.

* crlf -> lf

We do not want CRLF in the repo, let's replace this with LF.

* add newlines at eof

These are text-files, and should have a newline at EOF. Add the missing
ones.