Bring back styles cleanups (#138)
* remove needless timestamps

This is an attempt at cache-busting, but it's not really effective due
to needing manual update. Besides, this file hasn't been changed in at
least the last year.

We're currently hosting this file with a max-age of 604800, which is 7
days. We should probably consider lowering that somewhat, but for now
this should work.

* remove hard-coded content.pouet.net

* remove duplicate pngs

These are the same images, just not converted to SVG. Let's always use
the SVG version instead.

The difference between the a and b versions is just the brightness; we
can just fudge that with a css-filter to avoid having to create more
versions here, as discussed here:

https://www.pouet.net/topic.php?which=12585&page=1#c585781

* remove commented out styles

...and the corresponding assets.

* remove commented out style fragments

This just looks like leftover cruft from live-editing. Let's just nuke
it all.

* remove stale comment

I think this idea has proven itself, no need to be defensive or to place
blame.

* link to trumpet.svg in the style

* delete dead code

The red thread with these is that these are uses of POUET_CONTENT_URL
that are commented out. They provide no value, and I'd like to get rid
of the usage in the first place.

So let's remove these.

* re-export svgs as plain svg

This drops a bunch of needless Inskape meta-data.